Move data structure related code to own subdir
