+++ include/argent.h
@@ -36,9 +36,9 @@
 #include "./uuid.h"
 #include "./object.h"
 #include "./value.h"
-#include "./field.h"
-#include "./list.h"
-#include "./alist.h"
+#include "../src/ds/field.h"
+#include "../src/ds/list.h"
+#include "../src/ds/alist.h"
 #include "./plugin.h"
 #include "../src/test/test.h"
 #include "../src/http/http.h"
